add draft to vscdoe api in mrege
