+++ src/holesPanel/webview/vscodeApi.ts
@@ -1,4 +1,5 @@
-import { OutgoingMessage, IncomingMessage } from './messages';
+import { EffektHoleInfo } from '../effektHoleInfo';
+import { Location as LSPLocation } from 'vscode-languageserver-protocol';
 
 export interface VSCodeAPI {
   postMessage(msg: OutgoingMessage): void;
@@ -6,7 +7,16 @@ export interface VSCodeAPI {
 
 declare function acquireVsCodeApi(): VSCodeAPI;
 
-export const vscode = acquireVsCodeApi();
+export type OutgoingMessage =
+  | { command: 'jumpToHole'; holeId?: string }
+  | { command: 'jumpToDefinition'; definitionLocation: LSPLocation }
+  | { command: 'openCopilotChat'; holeId: string }
+  | { command: 'createDraft' };
 
-// Re-export types
-export type { OutgoingMessage, IncomingMessage };
+export type IncomingMessage =
+  | { command: 'highlightHole'; holeId: string }
+  | { command: 'updateHoles'; holes: EffektHoleInfo[] }
+  | { command: 'setShowHoles'; show: boolean }
+  | { command: 'focusPanel' };
+
+export const vscode = acquireVsCodeApi();
